feat(ThemeService): 이미 예약된 테마 삭제 불가

+++ src/main/java/roomescape/service/ThemeService.java
@@ -1,6 +1,7 @@
 package roomescape.service;
 
 import org.springframework.stereotype.Service;
+import roomescape.dao.ReservationDao;
 import roomescape.dao.ThemeDao;
 import roomescape.domain.Theme;
 import roomescape.dto.ThemeCreateRequest;
@@ -11,9 +12,11 @@ import java.util.List;
 public class ThemeService {
 
     private final ThemeDao themeDao;
+    private final ReservationDao reservationDao;
 
-    public ThemeService(ThemeDao themeDao) {
+    public ThemeService(ThemeDao themeDao, ReservationDao reservationDao) {
         this.themeDao = themeDao;
+        this.reservationDao = reservationDao;
     }
 
     public List<Theme> readThemes() {
@@ -28,6 +31,9 @@ public class ThemeService {
     }
 
     public void deleteTheme(Long id) {
+        if (reservationDao.isReservationsByThemeId(id)) {
+            throw new IllegalArgumentException("해당 테마를 사용하는 예약이 존재합니다.");
+        }
         themeDao.deleteTheme(id);
     }
 }
